requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
fastapi-cors==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-generativeai==0.3.2
lxml==4.9.3
gunicorn==21.2.0
structlog==23.2.0
healthcheck==1.3.3
slowapi==0.1.9
secure==0.3.0
